refactor: refactor the structure of column code
Refactor the structure of column code, make it dynamic

Refs:[]
